handle errors via err[2]
This inroduces a new technique of error handling in C libraries.
Should an error occur, information about the error is returned via
the `const char err[2]' parameter.  The first string is typically
a function name, and the second is a string which describes the
error.  Both strings normally come from the .rodata section.