redesigned NextMalloc/View in terms of MallocP
View() makes sense only for the zpkglist format - since it combines
four headers, which should be decompressed at once.  Thus, if the caller
wants just to "view" the header, this can save malloc+memcpy.  Other
formats need a buffer for each header which they malloc anyway.

Now, there are only two usage scenarios for the zpkglist format:
1) in apt; 2) with pkglist-query which builds contents_index_bin
in hasher.  In both cases, the header is loaded, and headerLoad()
takes ownership of the malloc'd blob.  So there's no use case
where View() without malloc might come useful.

Therefore, both Malloc() and View() can be implemented in terms
of MallocP() which provides the ability either to take ownership
of the malloc'd chunk or to leave it for the next iteration.

This change also removes the needMagic flag from the Next*() functions.
Roughly speaking, the magic is never needed.  In the usage scenarios
where headers are loaded, headerLoad() does not accept blobs with the
magic, and there is no way the magic can be skipped or discarded save
memmove().  In another usage scenario, which is genpkglist, I need
to recombine pkglist headers with headers read from rpms.  But then
again, nothing can make headerExport() to export magic.  So there's
no point in having some blobs with magic and some without.