README.md: introduced 128K+ jumbo frames
Compression with dictionary saves only a few hundred bytes per frame.
That's about 10% per average frame.  With 128K+ headers, however,
the savings are even less than average, because the dictionary effect
gets limited by the 64K window, so the dictionary data doesn't reach
to the second part of the header blob.

I expect the combined effect of reducing the size of jumbo frames from
256K to 128K and disabling dictionary compression to be less than 1%.
For comparison, each of the three steps: 1->2, 2->4 headers, dictionary,
improves compression by 10% or more (depending on the order in which
they are applied).  Compared to a solid LZ4 stream, this leaves
inefficiency of about 10%.  The next possible step, 4->8 headers,
reduces the inefficiency by one half.