removed lz4reader and support for .lz4 lists
Zstd is simply a superior technology, and can serve as a better
replacement for LZ4 in many cases.  The only area where LZ4 position
remains strong is client-side compression, possibly on a low-power CPU,
to exploit CPU/IO/virtual memory trade-off (this is what zpkglist files
are trying to achieve - to provide smooth access to a big repo on a
low-end machine).

I think the repo should provide xz and zstd-compressed lists.
There is no point in supporting anything else.