changed "char lead[16]" to "unsigned lead[4]"
This change also fixes the following code:

    headerDataSize(s->cur - 8);

Technically, (s->cur - 8) can be undefined behavior.  This is because
s->cur can point to the beginning of an mmap'd chunk, so after the
subtraction the pointer ends up outside of "array object" (C11 6.5.6.8).
Of course, this is a very Talmudic requirement that can hardly make any
difference in practice.