op-rpmheader.c: reimplemented OP(Bulk)
I want to implement "perfect concatenation", i.e. if a few frames
in different formats are concatenated in a single input, the magic
of a new frame should be identified just in time and the switch should
be completely seamless.  This isn't easy, though.  It took a lot of
brooding and ruminating.  For example, switching vtab inside vtab
member functions can have unexpected consequences.

Anyway, uncompressed streams were just slurped, but they need to be
parsed.  So this introduces a rather efficient implementation of
OP(Bulk) which parses the headers.  It is completely malloc-less,
returning the pointer to fda buffer.  It grabs as much as possible,
so it takes as few OP(Bulk) calls as possible.  Two places where
the switch can happen are marked with "XXX eos".