perfect concatenation works!
This introduces the new API function zpkglistRead() that provides
byte-stream reading; it is now used in zpkglistCompress(), which
no longer needs access to struct zpkglistReader internals.  An important
feature of zpkglistRead() is that it can read less bytes than requested
due to a frame boundary; the caller then should verify that a single
header doesn't span multiple frames.  I find it quite an elegant
solution, which pairs well with the "skewed reads" technique used
in zpkglistCompress().  It's not perfect, though: if the frame boundary
happens to be just right after the leading 16 bytes of a header,
the situation will go undetected.  But detecting this situation would
much complicate the API (something like "returns two zeroes in a row"
for "real EOF").  And designing the right kind of API is the hardest
part of the job, because it's about making the right compromises.