compress.c: combine zpkglistCompress into a single func with overloaded return
The whole point of zLoop function was to break out of the loop easily,
without having to clean up in each case.  Now, if one has to factor a
function just to break out of the loop easily, there must be something
a tad bit wrong with the language.  Or there has to be another way round.

So I find the technique of overloading the return operator very amusing.
(Just for the record, I am well aware that return is not an operator,
but rather a keyword and a statement.  I am also aware of gcc's
__attribute__((cleanup(func))).  The technique is quite amusing anyway.)