zreader.c: revamped zreader_begin
Now compiles better with the new spec and with the writing procedure.
In particular, partial leading frames are no longer permitted.  They are
not written either, to start with.  This and other changes are aimed at
better support for concatenation: the code should not assume that it
takes over the whole file.  Hence the sizes need stricter validation,
so that the boundaries can be established unambiguously.